dave_fetchAI/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
uagents==0.9.0
python-dotenv==1.0.0
asyncio==3.4.3 